Implemented .cpp template scripts and cleaned up Python C++ generation scripts.
Implemented the x86_data.template.cpp file so that it generates the
instruction data and writes it to the correct .cpp file

Got rid of old/unused code, (think) i normalised indentation for the
python).

Also added a bunch of comments highlighting the horrific HACKS that
exist in some of the templating code.

Also now the C++ generation step prints out some useful infomation
on which file is currently being generated, and when it's done, how long
it took. Useful for diagnosing long configure times/code generation
halts.