Resolve InstructionFrom by instruction size. Tests & fixes for WriteBuffer/ReadBuffer
X86Instruction::ResolveForm() will now choose the smallest of any given
instructions that have the same operands.
Added tests for above behaviour.

As a part of this instruction size is now calculated and added by the
X86 instruction C++ code generator.

Fixed bug in WriteBuffer: when creating WriteBuffer with a initial
capacity it would also initialise the size to that capacity whereas the
size _should_ be 0.
Added tests for this to prevent regressions.